requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
pymongo==4.14.1
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
google-auth>=2.20.0
